fix(io): restore pre-grouping atom order and hold the con id 64-bit wide
A .con file groups atoms by species because header lines 7 and 8 are a
type count and per-type counts, so a Matter whose species interleave is
written in one order and read back in another, while mode.dat and
direction.dat stay indexed in Matter order. The server writes both from
the same configuration, so the two sides have to undo the grouping the
same way. con2matter now sorts the frame's atoms by column 5.

readcon carries the id as a uint64_t and Matter::atomIndex held an int,
so an id past 2^31 lost its high half on the way in and came back out as
a different atom. The slot and both accessors are int64_t.

+++ client/Matter.cpp
@@ -200,7 +200,7 @@ void Matter::resize(const long int length) {
 
   atomIndex.resize(length);
   for (long i = 0; i < length; i++)
-    atomIndex(i) = static_cast<int>(i); // default: sequential
+    atomIndex(i) = static_cast<std::int64_t>(i); // default: sequential
   recomputePotential = true;
   recomputeMaskedForces = true;
   recomputeFreeMask = true;
